Improve and correct built-in symbols for Clojure
The current Clojure lexer uses an incomplete and buggy set of tokens. This
commit adds a more complete set of tokens that are either special forms or
part of clojure.core. It also refactors the symbols into categories based on
the groups defined in the [Glow][gh-g] syntax-highlighting library and uses
these categories to return an appropriate value to the caller of `#name_token`.

[gh-g]: https://github.com/venantius/glow